src/furk_my_mouse/requirements.txt
Pillow==11.1.0
pyautogui==0.9.54
pygame==2.6.1
